requirements.txt
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
ConfigArgParse==1.5.3
Faker==17.0.0
Flask==2.2.3
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==22.10.2
geventhttpclient==2.0.8
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.14.2
MarkupSafe==2.1.2
msgpack==1.0.4
psutil==5.9.4
python-dateutil==2.8.2
pyzmq==25.0.0
requests==2.28.2
roundrobin==0.0.4
six==1.16.0
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
zope.event==4.6
zope.interface==5.5.2

python-dotenv~=1.0.0